Crate shrev

Event channel, pull based, that use a ringbuffer for internal storage, to make it possible to do immutable reads.

See examples directory for examples.

Structs

EventChannel
The EventChannel, which is the central component of shrev.

EventIterator
Iterator over a slice of data in RingBufferStorage.

ReaderId
A reader ID which represents a subscription to the events pushed to the EventChannel.

Traits

Event
Marker trait for data to use with the EventChannel.